Force use of IPU for image conversion.
Force use of the scale plugin, and therefore the IPU, even if the
video output plugin indicates that it supports YUV formats. This is
necessary as the VPU output is in a non-standard YUV format that
libswscale cannot handle. This isn't an ideal solution and will be
improved later on.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>