Fix possible segfault in dmmu_match_user_mem_tlb()
This was introduced by commit db129a71. If the size passed to
dmmu_match_user_mem_tlb() is an exact multiple of the page size, the
code may segfault as it attempts to read off the end of the buffer.
Changed to round up the size given to the page size instead, which
should have the same effect intended by the original change, while
fixing the case of size being a multiple of PAGE_SIZE.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>